chore(builder): inline the error module into builder
The types are all generated from the builder module, so it makes no
sense at all to have them in a separate file, apart from being
annoying to work with.
